code refactor, separate http into multiple files
